pyproject.toml
[project]
name = "mlxgateway"
version = "0.1.0"
description = "MLX Gateway - OpenAI-compatible API server for Apple Silicon using MLX"
authors = [{ name = "yusufozgul" }]
requires-python = ">=3.11"
readme = "README.md"
license = "MIT"
keywords = ["mlx", "ai", "gateway", "openai", "apple-silicon"]
dependencies = [
    # Web Framework & API
    "fastapi>=0.128.0",
    "uvicorn>=0.40.0",
    "pydantic>=2.12.5",
    "python-multipart>=0.0.22",
    
    # MLX Libraries
    "mlx-lm==0.31.1",   # https://github.com/ml-explore/mlx-lm
    "mlx-vlm==0.4.0",  # https://github.com/Blaizzy/mlx-vlm
    "mlx-audio @ git+https://github.com/Blaizzy/mlx-audio.git",  # https://github.com/Blaizzy/mlx-audio (PyPI 0.4.0 pins mlx-lm==0.30.5, main branch fixed to 0.31.1)
    "mflux>=0.16.9",    # https://github.com/filipstrand/mflux 
    "huggingface-hub>=1.3.3",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    
    # Utilities
    "rich>=13.9.4",
    "setproctitle>=1.3.3",
]

[project.scripts]
mlxgateway = "mlxgateway.main:start"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true
